fix(AAiT-web_g3): move blog card to blog folder and update component imports
