+++ aait/web/group-3/starter_project_group3/components/blog/SmallBlogCard.tsx
